modify. 유저 권한설정 로직 변경 및 메시지리스너 위치 변경
