+++ src/hooks/useCallData.tsx
@@ -1,11 +1,9 @@
 import { onAuthStateChanged } from "@firebase/auth";
 import { FirestoreError, doc, updateDoc } from "@firebase/firestore";
-import { onMessage } from "@firebase/messaging";
 import { FirebaseError } from "@firebase/util";
 import { useAccountAction } from "@hooks/useAccount";
 import { useLoadingState } from "@hooks/useLoading";
 import { useTeamAction } from "@hooks/useTeam";
-import { useToastState } from "@hooks/useToast";
 import { Alert } from "@utils/Alert";
 import { fb } from "@utils/firebase";
 import dayjs from "dayjs";
@@ -16,7 +14,6 @@ export const useCallData = () => {
   const loadingState = useLoadingState();
   const accountState = useAccountAction();
   const teamState = useTeamAction();
-  const toastState = useToastState();
 
   // 데이터 불러오기
   const initData = async () => {
@@ -42,13 +39,16 @@ export const useCallData = () => {
         }
 
         let token: string | null = userData.token;
+        let isNotification: boolean = userData.isNotification;
         const permission = await fb.api.isNotification();
 
         // fcm 토큰 가져와서 유저정보 등록
         if (permission === "granted") {
           token = await fb.api.getFCMToken();
+          isNotification = true;
         }
 
+        // 유저 호출 쿼리
         const docRef = doc(fb.db, "accounts", userData.id);
 
         // 이메일 인증여부가 저장된 정보와 다른 경우
@@ -63,15 +63,11 @@ export const useCallData = () => {
           userData.token = token;
         }
 
-        // 메세지 리스너 등록
-        onMessage(fb.messaging, (payload) => {
-          if (payload.notification) {
-            const title = payload.notification.title ?? "";
-            const body = payload.notification.body ?? "";
-            toastState.toastQueue.enqueue({ title, body });
-            toastState.onRefresh();
-          }
-        });
+        // 알림 정보가 다른 경우
+        if (userData.isNotification !== isNotification) {
+          await updateDoc(docRef, { isNotification });
+          userData.isNotification = isNotification;
+        }
 
         // 팀 정보 불러오기
         const team = await fb.api.getTeamFromUserDocId(userData.id);
